backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-minio-backend==3.5.0
Pillow>=10.0.0
python-decouple==3.8 